+++ src/sysinfo.rs
@@ -77,7 +77,6 @@ impl SysInfo {
 //
 /// cargo watch -q -c -w src/ -x 'test sysinfo -- --test-threads=1 --nocapture'
 #[cfg(test)]
-#[allow(clippy::unwrap_used)]
 mod tests {
     use std::time::SystemTime;