+++ src/ws_messages/incoming.rs
@@ -49,7 +49,7 @@ pub fn to_struct(input: &str) -> Option<MessageValues> {
 ///
 /// cargo watch -q -c -w src/ -x 'test message_incoming -- --test-threads=1 --nocapture'
 #[cfg(test)]
-#[allow(clippy::unwrap_used)]
+#[expect(clippy::unwrap_used)]
 mod tests {
     use super::*;